+++ tests/test_models/__init__.py
@@ -1,0 +1,10 @@
+from .base import (check_norm_state, generate_backbone_demo_inputs,
+                   generate_gradcam_inputs, generate_recognizer_demo_inputs,
+                   get_audio_recognizer_cfg, get_cfg, get_localizer_cfg,
+                   get_recognizer_cfg)
+
+__all__ = [
+    'check_norm_state', 'generate_backbone_demo_inputs',
+    'generate_recognizer_demo_inputs', 'generate_gradcam_inputs', 'get_cfg',
+    'get_recognizer_cfg', 'get_audio_recognizer_cfg', 'get_localizer_cfg'
+]
