+++ tests/test_models/test_backbones.py
@@ -1,131 +1,15 @@
 import copy
 
-import numpy as np
 import pytest
 import torch
 import torch.nn as nn
 from mmcv.utils import _BatchNorm
 
-from mmaction.models import (C3D, X3D, ResNet, ResNet2Plus1d, ResNet3d,
-                             ResNet3dCSN, ResNet3dLayer, ResNet3dSlowFast,
-                             ResNet3dSlowOnly, ResNetAudio, ResNetTIN,
-                             ResNetTSM)
+from mmaction.models import (C3D, X3D, ResNet2Plus1d, ResNet3dCSN,
+                             ResNet3dSlowFast, ResNet3dSlowOnly, ResNetAudio,
+                             ResNetTIN, ResNetTSM)
 from mmaction.models.backbones.resnet_tsm import NL3DWrapper
-
-
-def check_norm_state(modules, train_state):
-    """Check if norm layer is in correct train state."""
-    for mod in modules:
-        if isinstance(mod, _BatchNorm):
-            if mod.training != train_state:
-                return False
-    return True
-
-
-def test_resnet_backbone():
-    """Test resnet backbone."""
-    with pytest.raises(KeyError):
-        # ResNet depth should be in [18, 34, 50, 101, 152]
-        ResNet(20)
-
-    with pytest.raises(AssertionError):
-        # In ResNet: 1 <= num_stages <= 4
-        ResNet(50, num_stages=0)
-
-    with pytest.raises(AssertionError):
-        # In ResNet: 1 <= num_stages <= 4
-        ResNet(50, num_stages=5)
-
-    with pytest.raises(AssertionError):
-        # len(strides) == len(dilations) == num_stages
-        ResNet(50, strides=(1, ), dilations=(1, 1), num_stages=3)
-
-    with pytest.raises(TypeError):
-        # pretrain must be a str
-        resnet50 = ResNet(50, pretrained=0)
-        resnet50.init_weights()
-
-    with pytest.raises(AssertionError):
-        # style must be in ['pytorch', 'caffe']
-        ResNet(18, style='tensorflow')
-
-    with pytest.raises(AssertionError):
-        # assert not with_cp
-        ResNet(18, with_cp=True)
-
-    # resnet with depth 18, norm_eval False, initial weights
-    resnet18 = ResNet(18)
-    resnet18.init_weights()
-
-    # resnet with depth 50, norm_eval True
-    resnet50 = ResNet(50, norm_eval=True)
-    resnet50.init_weights()
-    resnet50.train()
-    assert check_norm_state(resnet50.modules(), False)
-
-    # resnet with depth 50, norm_eval True, pretrained
-    resnet50_pretrain = ResNet(
-        pretrained='torchvision://resnet50', depth=50, norm_eval=True)
-    resnet50_pretrain.init_weights()
-    resnet50_pretrain.train()
-    assert check_norm_state(resnet50_pretrain.modules(), False)
-
-    # resnet with depth 50, norm_eval True, frozen_stages 1
-    frozen_stages = 1
-    resnet50_frozen = ResNet(50, frozen_stages=frozen_stages)
-    resnet50_frozen.init_weights()
-    resnet50_frozen.train()
-    assert resnet50_frozen.conv1.bn.training is False
-    for layer in resnet50_frozen.conv1.modules():
-        for param in layer.parameters():
-            assert param.requires_grad is False
-    for i in range(1, frozen_stages + 1):
-        layer = getattr(resnet50_frozen, f'layer{i}')
-        for mod in layer.modules():
-            if isinstance(mod, _BatchNorm):
-                assert mod.training is False
-        for param in layer.parameters():
-            assert param.requires_grad is False
-
-    # resnet with depth 50, partial batchnorm
-    resnet_pbn = ResNet(50, partial_bn=True)
-    resnet_pbn.train()
-    count_bn = 0
-    for m in resnet_pbn.modules():
-        if isinstance(m, nn.BatchNorm2d):
-            count_bn += 1
-            if count_bn >= 2:
-                assert m.weight.requires_grad is False
-                assert m.bias.requires_grad is False
-                assert m.training is False
-            else:
-                assert m.weight.requires_grad is True
-                assert m.bias.requires_grad is True
-                assert m.training is True
-
-    input_shape = (1, 3, 64, 64)
-    imgs = _demo_inputs(input_shape)
-
-    # resnet with depth 18 inference
-    resnet18 = ResNet(18, norm_eval=False)
-    resnet18.init_weights()
-    resnet18.train()
-    feat = resnet18(imgs)
-    assert feat.shape == torch.Size([1, 512, 2, 2])
-
-    # resnet with depth 50 inference
-    resnet50 = ResNet(50, norm_eval=False)
-    resnet50.init_weights()
-    resnet50.train()
-    feat = resnet50(imgs)
-    assert feat.shape == torch.Size([1, 2048, 2, 2])
-
-    # resnet with depth 50 in caffe style inference
-    resnet50_caffe = ResNet(50, style='caffe', norm_eval=False)
-    resnet50_caffe.init_weights()
-    resnet50_caffe.train()
-    feat = resnet50_caffe(imgs)
-    assert feat.shape == torch.Size([1, 2048, 2, 2])
+from .base import check_norm_state, generate_backbone_demo_inputs
 
 
 def test_x3d_backbone():
@@ -217,7 +101,7 @@ def test_x3d_backbone():
 
     # x3d_s inference
     input_shape = (1, 3, 13, 160, 160)
-    imgs = _demo_inputs(input_shape)
+    imgs = generate_backbone_demo_inputs(input_shape)
     # parrots 3dconv is only implemented on gpu
     if torch.__version__ == 'parrots':
         if torch.cuda.is_available():
@@ -231,7 +115,7 @@ def test_x3d_backbone():
 
     # x3d_m inference
     input_shape = (1, 3, 16, 224, 224)
-    imgs = _demo_inputs(input_shape)
+    imgs = generate_backbone_demo_inputs(input_shape)
     # parrots 3dconv is only implemented on gpu
     if torch.__version__ == 'parrots':
         if torch.cuda.is_available():
@@ -242,333 +126,6 @@ def test_x3d_backbone():
     else:
         feat = x3d_s_frozen(imgs)
         assert feat.shape == torch.Size([1, 432, 16, 7, 7])
-
-
-def test_resnet3d_backbone():
-    """Test resnet3d backbone."""
-    with pytest.raises(AssertionError):
-        # In ResNet3d: 1 <= num_stages <= 4
-        ResNet3d(34, None, num_stages=0)
-
-    with pytest.raises(AssertionError):
-        # In ResNet3d: 1 <= num_stages <= 4
-        ResNet3d(34, None, num_stages=5)
-
-    with pytest.raises(AssertionError):
-        # In ResNet3d: 1 <= num_stages <= 4
-        ResNet3d(50, None, num_stages=0)
-
-    with pytest.raises(AssertionError):
-        # In ResNet3d: 1 <= num_stages <= 4
-        ResNet3d(50, None, num_stages=5)
-
-    with pytest.raises(AssertionError):
-        # len(spatial_strides) == len(temporal_strides)
-        # == len(dilations) == num_stages
-        ResNet3d(
-            50,
-            None,
-            spatial_strides=(1, ),
-            temporal_strides=(1, 1),
-            dilations=(1, 1, 1),
-            num_stages=4)
-
-    with pytest.raises(AssertionError):
-        # len(spatial_strides) == len(temporal_strides)
-        # == len(dilations) == num_stages
-        ResNet3d(
-            34,
-            None,
-            spatial_strides=(1, ),
-            temporal_strides=(1, 1),
-            dilations=(1, 1, 1),
-            num_stages=4)
-
-    with pytest.raises(TypeError):
-        # pretrain must be str or None.
-        resnet3d_34 = ResNet3d(34, ['resnet', 'bninception'])
-        resnet3d_34.init_weights()
-
-    with pytest.raises(TypeError):
-        # pretrain must be str or None.
-        resnet3d_50 = ResNet3d(50, ['resnet', 'bninception'])
-        resnet3d_50.init_weights()
-
-    # resnet3d with depth 34, no pretrained, norm_eval True
-    resnet3d_34 = ResNet3d(34, None, pretrained2d=False, norm_eval=True)
-    resnet3d_34.init_weights()
-    resnet3d_34.train()
-    assert check_norm_state(resnet3d_34.modules(), False)
-
-    # resnet3d with depth 50, no pretrained, norm_eval True
-    resnet3d_50 = ResNet3d(50, None, pretrained2d=False, norm_eval=True)
-    resnet3d_50.init_weights()
-    resnet3d_50.train()
-    assert check_norm_state(resnet3d_50.modules(), False)
-
-    # resnet3d with depth 50, pretrained2d, norm_eval True
-    resnet3d_50_pretrain = ResNet3d(
-        50, 'torchvision://resnet50', norm_eval=True)
-    resnet3d_50_pretrain.init_weights()
-    resnet3d_50_pretrain.train()
-    assert check_norm_state(resnet3d_50_pretrain.modules(), False)
-    from mmcv.runner import _load_checkpoint
-    chkp_2d = _load_checkpoint('torchvision://resnet50')
-    for name, module in resnet3d_50_pretrain.named_modules():
-        if len(name.split('.')) == 4:
-            # layer.block.module.submodule
-            prefix = name.split('.')[:2]
-            module_type = name.split('.')[2]
-            submodule_type = name.split('.')[3]
-
-            if module_type == 'downsample':
-                name2d = name.replace('conv', '0').replace('bn', '1')
-            else:
-                layer_id = name.split('.')[2][-1]
-                name2d = prefix[0] + '.' + prefix[1] + '.' + \
-                    submodule_type + layer_id
-
-            if isinstance(module, nn.Conv3d):
-                conv2d_weight = chkp_2d[name2d + '.weight']
-                conv3d_weight = getattr(module, 'weight').data
-                assert torch.equal(
-                    conv3d_weight,
-                    conv2d_weight.data.unsqueeze(2).expand_as(conv3d_weight) /
-                    conv3d_weight.shape[2])
-                if getattr(module, 'bias') is not None:
-                    conv2d_bias = chkp_2d[name2d + '.bias']
-                    conv3d_bias = getattr(module, 'bias').data
-                    assert torch.equal(conv2d_bias, conv3d_bias)
-
-            elif isinstance(module, nn.BatchNorm3d):
-                for pname in ['weight', 'bias', 'running_mean', 'running_var']:
-                    param_2d = chkp_2d[name2d + '.' + pname]
-                    param_3d = getattr(module, pname).data
-                assert torch.equal(param_2d, param_3d)
-
-    conv3d = resnet3d_50_pretrain.conv1.conv
-    assert torch.equal(
-        conv3d.weight,
-        chkp_2d['conv1.weight'].unsqueeze(2).expand_as(conv3d.weight) /
-        conv3d.weight.shape[2])
-    conv3d = resnet3d_50_pretrain.layer3[2].conv2.conv
-    assert torch.equal(
-        conv3d.weight, chkp_2d['layer3.2.conv2.weight'].unsqueeze(2).expand_as(
-            conv3d.weight) / conv3d.weight.shape[2])
-
-    # resnet3d with depth 34, no pretrained, norm_eval False
-    resnet3d_34_no_bn_eval = ResNet3d(
-        34, None, pretrained2d=False, norm_eval=False)
-    resnet3d_34_no_bn_eval.init_weights()
-    resnet3d_34_no_bn_eval.train()
-    assert check_norm_state(resnet3d_34_no_bn_eval.modules(), True)
-
-    # resnet3d with depth 50, no pretrained, norm_eval False
-    resnet3d_50_no_bn_eval = ResNet3d(
-        50, None, pretrained2d=False, norm_eval=False)
-    resnet3d_50_no_bn_eval.init_weights()
-    resnet3d_50_no_bn_eval.train()
-    assert check_norm_state(resnet3d_50_no_bn_eval.modules(), True)
-
-    # resnet3d with depth 34, no pretrained, frozen_stages, norm_eval False
-    frozen_stages = 1
-    resnet3d_34_frozen = ResNet3d(
-        34, None, pretrained2d=False, frozen_stages=frozen_stages)
-    resnet3d_34_frozen.init_weights()
-    resnet3d_34_frozen.train()
-    assert resnet3d_34_frozen.conv1.bn.training is False
-    for param in resnet3d_34_frozen.conv1.parameters():
-        assert param.requires_grad is False
-    for i in range(1, frozen_stages + 1):
-        layer = getattr(resnet3d_34_frozen, f'layer{i}')
-        for mod in layer.modules():
-            if isinstance(mod, _BatchNorm):
-                assert mod.training is False
-        for param in layer.parameters():
-            assert param.requires_grad is False
-    # test zero_init_residual
-    for m in resnet3d_34_frozen.modules():
-        if hasattr(m, 'conv2'):
-            assert torch.equal(m.conv2.bn.weight,
-                               torch.zeros_like(m.conv2.bn.weight))
-            assert torch.equal(m.conv2.bn.bias,
-                               torch.zeros_like(m.conv2.bn.bias))
-
-    # resnet3d with depth 50, no pretrained, frozen_stages, norm_eval False
-    frozen_stages = 1
-    resnet3d_50_frozen = ResNet3d(
-        50, None, pretrained2d=False, frozen_stages=frozen_stages)
-    resnet3d_50_frozen.init_weights()
-    resnet3d_50_frozen.train()
-    assert resnet3d_50_frozen.conv1.bn.training is False
-    for param in resnet3d_50_frozen.conv1.parameters():
-        assert param.requires_grad is False
-    for i in range(1, frozen_stages + 1):
-        layer = getattr(resnet3d_50_frozen, f'layer{i}')
-        for mod in layer.modules():
-            if isinstance(mod, _BatchNorm):
-                assert mod.training is False
-        for param in layer.parameters():
-            assert param.requires_grad is False
-    # test zero_init_residual
-    for m in resnet3d_50_frozen.modules():
-        if hasattr(m, 'conv3'):
-            assert torch.equal(m.conv3.bn.weight,
-                               torch.zeros_like(m.conv3.bn.weight))
-            assert torch.equal(m.conv3.bn.bias,
-                               torch.zeros_like(m.conv3.bn.bias))
-
-    # resnet3d frozen with depth 34 inference
-    input_shape = (1, 3, 6, 64, 64)
-    imgs = _demo_inputs(input_shape)
-    # parrots 3dconv is only implemented on gpu
-    if torch.__version__ == 'parrots':
-        if torch.cuda.is_available():
-            resnet3d_34_frozen = resnet3d_34_frozen.cuda()
-            imgs_gpu = imgs.cuda()
-            feat = resnet3d_34_frozen(imgs_gpu)
-            assert feat.shape == torch.Size([1, 512, 1, 2, 2])
-    else:
-        feat = resnet3d_34_frozen(imgs)
-        assert feat.shape == torch.Size([1, 512, 1, 2, 2])
-
-    # resnet3d with depth 50 inference
-    input_shape = (1, 3, 6, 64, 64)
-    imgs = _demo_inputs(input_shape)
-    # parrots 3dconv is only implemented on gpu
-    if torch.__version__ == 'parrots':
-        if torch.cuda.is_available():
-            resnet3d_50_frozen = resnet3d_50_frozen.cuda()
-            imgs_gpu = imgs.cuda()
-            feat = resnet3d_50_frozen(imgs_gpu)
-            assert feat.shape == torch.Size([1, 2048, 1, 2, 2])
-    else:
-        feat = resnet3d_50_frozen(imgs)
-        assert feat.shape == torch.Size([1, 2048, 1, 2, 2])
-
-    # resnet3d with depth 50 in caffe style inference
-    resnet3d_50_caffe = ResNet3d(50, None, pretrained2d=False, style='caffe')
-    resnet3d_50_caffe.init_weights()
-    resnet3d_50_caffe.train()
-
-    # parrots 3dconv is only implemented on gpu
-    if torch.__version__ == 'parrots':
-        if torch.cuda.is_available():
-            resnet3d_50_caffe = resnet3d_50_caffe.cuda()
-            imgs_gpu = imgs.cuda()
-            feat = resnet3d_50_caffe(imgs_gpu)
-            assert feat.shape == torch.Size([1, 2048, 1, 2, 2])
-    else:
-        feat = resnet3d_50_caffe(imgs)
-        assert feat.shape == torch.Size([1, 2048, 1, 2, 2])
-
-    # resnet3d with depth 34 in caffe style inference
-    resnet3d_34_caffe = ResNet3d(34, None, pretrained2d=False, style='caffe')
-    resnet3d_34_caffe.init_weights()
-    resnet3d_34_caffe.train()
-    # parrots 3dconv is only implemented on gpu
-    if torch.__version__ == 'parrots':
-        if torch.cuda.is_available():
-            resnet3d_34_caffe = resnet3d_34_caffe.cuda()
-            imgs_gpu = imgs.cuda()
-            feat = resnet3d_34_caffe(imgs_gpu)
-            assert feat.shape == torch.Size([1, 512, 1, 2, 2])
-    else:
-        feat = resnet3d_34_caffe(imgs)
-        assert feat.shape == torch.Size([1, 512, 1, 2, 2])
-
-    # resnet3d with depth with 3x3x3 inflate_style inference
-    resnet3d_50_1x1x1 = ResNet3d(
-        50, None, pretrained2d=False, inflate_style='3x3x3')
-    resnet3d_50_1x1x1.init_weights()
-    resnet3d_50_1x1x1.train()
-    # parrots 3dconv is only implemented on gpu
-    if torch.__version__ == 'parrots':
-        if torch.cuda.is_available():
-            resnet3d_50_1x1x1 = resnet3d_50_1x1x1.cuda()
-            imgs_gpu = imgs.cuda()
-            feat = resnet3d_50_1x1x1(imgs_gpu)
-            assert feat.shape == torch.Size([1, 2048, 1, 2, 2])
-    else:
-        feat = resnet3d_50_1x1x1(imgs)
-        assert feat.shape == torch.Size([1, 2048, 1, 2, 2])
-
-    resnet3d_34_1x1x1 = ResNet3d(
-        34, None, pretrained2d=False, inflate_style='3x3x3')
-    resnet3d_34_1x1x1.init_weights()
-    resnet3d_34_1x1x1.train()
-
-    # parrots 3dconv is only implemented on gpu
-    if torch.__version__ == 'parrots':
-        if torch.cuda.is_available():
-            resnet3d_34_1x1x1 = resnet3d_34_1x1x1.cuda()
-            imgs_gpu = imgs.cuda()
-            feat = resnet3d_34_1x1x1(imgs_gpu)
-            assert feat.shape == torch.Size([1, 512, 1, 2, 2])
-    else:
-        feat = resnet3d_34_1x1x1(imgs)
-        assert feat.shape == torch.Size([1, 512, 1, 2, 2])
-
-    # resnet3d with non-local module
-    non_local_cfg = dict(
-        sub_sample=True,
-        use_scale=False,
-        norm_cfg=dict(type='BN3d', requires_grad=True),
-        mode='embedded_gaussian')
-    non_local = ((0, 0, 0), (1, 0, 1, 0), (1, 0, 1, 0, 1, 0), (0, 0, 0))
-    resnet3d_nonlocal = ResNet3d(
-        50,
-        None,
-        pretrained2d=False,
-        non_local=non_local,
-        non_local_cfg=non_local_cfg)
-    resnet3d_nonlocal.init_weights()
-    for layer_name in ['layer2', 'layer3']:
-        layer = getattr(resnet3d_nonlocal, layer_name)
-        for i, _ in enumerate(layer):
-            if i % 2 == 0:
-                assert hasattr(layer[i], 'non_local_block')
-
-    feat = resnet3d_nonlocal(imgs)
-    assert feat.shape == torch.Size([1, 2048, 1, 2, 2])
-
-
-def test_resnet3d_layer():
-    with pytest.raises(AssertionError):
-        ResNet3dLayer(22, None)
-
-    with pytest.raises(AssertionError):
-        ResNet3dLayer(50, None, stage=4)
-
-    res_layer = ResNet3dLayer(50, None, stage=3, norm_eval=True)
-    res_layer.init_weights()
-    res_layer.train()
-    input_shape = (1, 1024, 1, 4, 4)
-    imgs = _demo_inputs(input_shape)
-    if torch.__version__ == 'parrots':
-        if torch.cuda.is_available():
-            res_layer = res_layer.cuda()
-            imgs_gpu = imgs.cuda()
-            feat = res_layer(imgs_gpu)
-            assert feat.shape == torch.Size([1, 2048, 1, 2, 2])
-    else:
-        feat = res_layer(imgs)
-        assert feat.shape == torch.Size([1, 2048, 1, 2, 2])
-
-    res_layer = ResNet3dLayer(
-        50, 'torchvision://resnet50', stage=3, all_frozen=True)
-    res_layer.init_weights()
-    res_layer.train()
-    imgs = _demo_inputs(input_shape)
-    if torch.__version__ == 'parrots':
-        if torch.cuda.is_available():
-            res_layer = res_layer.cuda()
-            imgs_gpu = imgs.cuda()
-            feat = res_layer(imgs_gpu)
-            assert feat.shape == torch.Size([1, 2048, 1, 2, 2])
-    else:
-        feat = res_layer(imgs)
-        assert feat.shape == torch.Size([1, 2048, 1, 2, 2])
 
 
 def test_resnet2plus1d_backbone():
@@ -609,7 +166,7 @@ def test_resnet2plus1d_backbone():
         for param in layer.parameters():
             assert param.requires_grad is False
     input_shape = (1, 3, 8, 64, 64)
-    imgs = _demo_inputs(input_shape)
+    imgs = generate_backbone_demo_inputs(input_shape)
     # parrots 3dconv is only implemented on gpu
     if torch.__version__ == 'parrots':
         if torch.cuda.is_available():
@@ -648,7 +205,7 @@ def test_resnet2plus1d_backbone():
         for param in layer.parameters():
             assert param.requires_grad is False
     input_shape = (1, 3, 8, 64, 64)
-    imgs = _demo_inputs(input_shape)
+    imgs = generate_backbone_demo_inputs(input_shape)
 
     # parrots 3dconv is only implemented on gpu
     if torch.__version__ == 'parrots':
@@ -673,7 +230,7 @@ def test_resnet_tsm_backbone():
     from mmaction.models.backbones.resnet_tsm import TemporalShift
 
     input_shape = (8, 3, 64, 64)
-    imgs = _demo_inputs(input_shape)
+    imgs = generate_backbone_demo_inputs(input_shape)
 
     # resnet_tsm with depth 50
     resnet_tsm_50 = ResNetTSM(50)
@@ -760,7 +317,7 @@ def test_resnet_tsm_backbone():
 
     # TSM with temporal pool + non-local forward
     input_shape = (16, 3, 32, 32)
-    imgs = _demo_inputs(input_shape)
+    imgs = generate_backbone_demo_inputs(input_shape)
     feat = resnet_tsm_50_full(imgs)
     assert feat.shape == torch.Size([8, 2048, 1, 1])
 
@@ -813,7 +370,7 @@ def test_slowfast_backbone():
 
     # slowfast w/o lateral connection inference test
     input_shape = (1, 3, 8, 64, 64)
-    imgs = _demo_inputs(input_shape)
+    imgs = generate_backbone_demo_inputs(input_shape)
     # parrots 3dconv is only implemented on gpu
     if torch.__version__ == 'parrots':
         if torch.cuda.is_available():
@@ -868,7 +425,7 @@ def test_slowfast_backbone():
 
     # slowfast inference test
     input_shape = (1, 3, 8, 64, 64)
-    imgs = _demo_inputs(input_shape)
+    imgs = generate_backbone_demo_inputs(input_shape)
     # parrots 3dconv is only implemented on gpu
     if torch.__version__ == 'parrots':
         if torch.cuda.is_available():
@@ -896,7 +453,7 @@ def test_slowonly_backbone():
 
     # SlowOnly inference test
     input_shape = (1, 3, 8, 64, 64)
-    imgs = _demo_inputs(input_shape)
+    imgs = generate_backbone_demo_inputs(input_shape)
     # parrots 3dconv is only implemented on gpu
     if torch.__version__ == 'parrots':
         if torch.cuda.is_available():
@@ -915,7 +472,7 @@ def test_resnet_csn_backbone():
         ResNet3dCSN(152, None, bottleneck_mode='id')
 
     input_shape = (2, 3, 6, 64, 64)
-    imgs = _demo_inputs(input_shape)
+    imgs = generate_backbone_demo_inputs(input_shape)
 
     resnet3d_csn_frozen = ResNet3dCSN(
         152, None, bn_frozen=True, norm_eval=True)
@@ -980,7 +537,7 @@ def test_resnet_csn_backbone():
 def test_c3d_backbone():
     """Test c3d backbone."""
     input_shape = (1, 3, 16, 112, 112)
-    imgs = _demo_inputs(input_shape)
+    imgs = generate_backbone_demo_inputs(input_shape)
 
     # c3d inference test
     c3d = C3D()
@@ -1000,7 +557,7 @@ def test_c3d_backbone():
 def test_resnet_audio_backbone():
     """Test ResNetAudio backbone."""
     input_shape = (1, 1, 16, 16)
-    spec = _demo_inputs(input_shape)
+    spec = generate_backbone_demo_inputs(input_shape)
     # inference
     audioonly = ResNetAudio(50, None)
     audioonly.init_weights()
@@ -1051,22 +608,9 @@ def test_resnet_tin_backbone():
                 assert m.bias.requires_grad is True
 
     input_shape = (8, 3, 64, 64)
-    imgs = _demo_inputs(input_shape).cuda()
+    imgs = generate_backbone_demo_inputs(input_shape).cuda()
     resnet_tin = resnet_tin.cuda()
 
     # resnet_tin with normal cfg inference
     feat = resnet_tin(imgs)
     assert feat.shape == torch.Size([8, 2048, 2, 2])
-
-
-def _demo_inputs(input_shape=(1, 3, 64, 64)):
-    """Create a superset of inputs needed to run backbone.
-
-    Args:
-        input_shape (tuple): input batch dimensions.
-            Default: (1, 3, 64, 64).
-    """
-    imgs = np.random.random(input_shape)
-    imgs = torch.FloatTensor(imgs)
-
-    return imgs
