+++ tests/test_models/test_gradcam.py
@@ -1,27 +1,9 @@
-import os.path as osp
-
-import mmcv
-import numpy as np
 import pytest
 import torch
 
 from mmaction.models import build_recognizer
 from mmaction.utils.gradcam_utils import GradCAM
-
-
-def _get_cfg(fname):
-    """Grab configs necessary to create a recognizer.
-
-    These are deep copied to allow for safe modification of parameters without
-    influencing other tests.
-    """
-    repo_dpath = osp.dirname(osp.dirname(__file__))
-    config_dpath = osp.join(repo_dpath, 'configs/recognition')
-    config_fpath = osp.join(config_dpath, fname)
-    if not osp.exists(config_dpath):
-        raise Exception('Cannot find config path')
-    config = mmcv.Config.fromfile(config_fpath)
-    return config
+from .base import generate_gradcam_inputs, get_cfg
 
 
 def _get_target_shapes(input_shape, num_classes=400, model_type='2D'):
@@ -45,37 +27,12 @@ def _get_target_shapes(input_shape, num_classes=400, model_type='2D'):
     return blended_imgs_target_shape, preds_target_shape
 
 
-def _generate_gradcam_inputs(input_shape=(1, 3, 3, 224, 224), model_type='2D'):
-    """Create a superset of inputs needed to run gradcam.
-
-    Args:
-        input_shape (tuple[int]): input batch dimensions.
-            Default: (1, 3, 3, 224, 224).
-        model_type (str): Model type for data generation, from {'2D', '3D'}.
-            Default:'2D'
-    return:
-        dict: model inputs, including two keys, ``imgs`` and ``label``.
-    """
-    imgs = np.random.random(input_shape)
-
-    if model_type in ['2D', '3D']:
-        gt_labels = torch.LongTensor([2] * input_shape[0])
-    else:
-        raise ValueError(f'Data type {model_type} is not available')
-
-    inputs = {
-        'imgs': torch.FloatTensor(imgs),
-        'label': gt_labels,
-    }
-    return inputs
-
-
 def _do_test_2D_models(recognizer,
                        target_layer_name,
                        input_shape,
                        num_classes=400,
                        device='cpu'):
-    demo_inputs = _generate_gradcam_inputs(input_shape)
+    demo_inputs = generate_gradcam_inputs(input_shape)
     demo_inputs['imgs'] = demo_inputs['imgs'].to(device)
     demo_inputs['label'] = demo_inputs['label'].to(device)
 
@@ -100,7 +57,7 @@ def _do_test_3D_models(recognizer,
                        num_classes=400):
     blended_imgs_target_shape, preds_target_shape = _get_target_shapes(
         input_shape, num_classes=num_classes, model_type='3D')
-    demo_inputs = _generate_gradcam_inputs(input_shape, '3D')
+    demo_inputs = generate_gradcam_inputs(input_shape, '3D')
 
     # parrots 3dconv is only implemented on gpu
     if torch.__version__ == 'parrots':
@@ -130,7 +87,8 @@ def _do_test_3D_models(recognizer,
 
 
 def test_tsn():
-    config = _get_cfg('tsn/tsn_r50_1x1x3_100e_kinetics400_rgb.py')
+    config = get_cfg('recognition',
+                     'tsn/tsn_r50_1x1x3_100e_kinetics400_rgb.py')
     config.model['backbone']['pretrained'] = None
     recognizer = build_recognizer(config.model, test_cfg=config.test_cfg)
     recognizer.cfg = config
@@ -142,7 +100,8 @@ def test_tsn():
 
 
 def test_i3d():
-    config = _get_cfg('i3d/i3d_r50_32x2x1_100e_kinetics400_rgb.py')
+    config = get_cfg('recognition',
+                     'i3d/i3d_r50_32x2x1_100e_kinetics400_rgb.py')
     config.model['backbone']['pretrained2d'] = False
     config.model['backbone']['pretrained'] = None
 
@@ -156,7 +115,8 @@ def test_i3d():
 
 
 def test_r2plus1d():
-    config = _get_cfg('r2plus1d/r2plus1d_r34_8x8x1_180e_kinetics400_rgb.py')
+    config = get_cfg('recognition',
+                     'r2plus1d/r2plus1d_r34_8x8x1_180e_kinetics400_rgb.py')
     config.model['backbone']['pretrained2d'] = False
     config.model['backbone']['pretrained'] = None
     config.model['backbone']['norm_cfg'] = dict(type='BN3d')
@@ -171,7 +131,8 @@ def test_r2plus1d():
 
 
 def test_slowfast():
-    config = _get_cfg('slowfast/slowfast_r50_4x16x1_256e_kinetics400_rgb.py')
+    config = get_cfg('recognition',
+                     'slowfast/slowfast_r50_4x16x1_256e_kinetics400_rgb.py')
 
     recognizer = build_recognizer(config.model, test_cfg=config.test_cfg)
     recognizer.cfg = config
@@ -183,7 +144,7 @@ def test_slowfast():
 
 
 def test_tsm():
-    config = _get_cfg('tsm/tsm_r50_1x1x8_50e_kinetics400_rgb.py')
+    config = get_cfg('recognition', 'tsm/tsm_r50_1x1x8_50e_kinetics400_rgb.py')
     config.model['backbone']['pretrained'] = None
     target_layer_name = 'backbone/layer4/1/relu'
 
@@ -202,7 +163,8 @@ def test_tsm():
 
 
 def test_csn():
-    config = _get_cfg(
+    config = get_cfg(
+        'recognition',
         'csn/ircsn_ig65m_pretrained_r152_32x2x1_58e_kinetics400_rgb.py')
     config.model['backbone']['pretrained2d'] = False
     config.model['backbone']['pretrained'] = None
@@ -218,7 +180,7 @@ def test_csn():
 def test_tpn():
     target_layer_name = 'backbone/layer4/1/relu'
 
-    config = _get_cfg('tpn/tpn_tsm_r50_1x1x8_150e_sthv1_rgb.py')
+    config = get_cfg('recognition', 'tpn/tpn_tsm_r50_1x1x8_150e_sthv1_rgb.py')
     config.model['backbone']['pretrained'] = None
     recognizer = build_recognizer(config.model, test_cfg=config.test_cfg)
     recognizer.cfg = config
@@ -226,7 +188,8 @@ def test_tpn():
     input_shape = (1, 8, 3, 32, 32)
     _do_test_2D_models(recognizer, target_layer_name, input_shape, 174)
 
-    config = _get_cfg('tpn/tpn_slowonly_r50_8x8x1_150e_kinetics_rgb.py')
+    config = get_cfg('recognition',
+                     'tpn/tpn_slowonly_r50_8x8x1_150e_kinetics_rgb.py')
     config.model['backbone']['pretrained'] = None
     recognizer = build_recognizer(config.model, test_cfg=config.test_cfg)
     recognizer.cfg = config
@@ -235,7 +198,8 @@ def test_tpn():
 
 
 def test_c3d():
-    config = _get_cfg('c3d/c3d_sports1m_16x1x1_45e_ucf101_rgb.py')
+    config = get_cfg('recognition',
+                     'c3d/c3d_sports1m_16x1x1_45e_ucf101_rgb.py')
     config.model['backbone']['pretrained'] = None
     recognizer = build_recognizer(config.model, test_cfg=config.test_cfg)
     recognizer.cfg = config
@@ -247,7 +211,8 @@ def test_c3d():
 @pytest.mark.skipif(
     not torch.cuda.is_available(), reason='requires CUDA support')
 def test_tin():
-    config = _get_cfg('tin/tin_tsm_finetune_r50_1x1x8_50e_kinetics400_rgb.py')
+    config = get_cfg('recognition',
+                     'tin/tin_tsm_finetune_r50_1x1x8_50e_kinetics400_rgb.py')
     config.model['backbone']['pretrained'] = None
     target_layer_name = 'backbone/layer4/1/relu'
 
@@ -259,7 +224,8 @@ def test_tin():
 
 
 def test_x3d():
-    config = _get_cfg('x3d/x3d_s_13x6x1_facebook_kinetics400_rgb.py')
+    config = get_cfg('recognition',
+                     'x3d/x3d_s_13x6x1_facebook_kinetics400_rgb.py')
     config.model['backbone']['pretrained'] = None
     recognizer = build_recognizer(config.model, test_cfg=config.test_cfg)
     recognizer.cfg = config
