+++ tests/test_models/test_recognizers/test_recognizer2d.py
@@ -1,0 +1,77 @@
+import torch
+
+from mmaction.models import build_recognizer
+from ..base import generate_recognizer_demo_inputs, get_recognizer_cfg
+
+
+def test_tsn():
+    model, train_cfg, test_cfg = get_recognizer_cfg(
+        'tsn/tsn_r50_1x1x3_100e_kinetics400_rgb.py')
+    model['backbone']['pretrained'] = None
+
+    recognizer = build_recognizer(
+        model, train_cfg=train_cfg, test_cfg=test_cfg)
+
+    input_shape = (1, 3, 3, 32, 32)
+    demo_inputs = generate_recognizer_demo_inputs(input_shape)
+
+    imgs = demo_inputs['imgs']
+    gt_labels = demo_inputs['gt_labels']
+
+    losses = recognizer(imgs, gt_labels)
+    assert isinstance(losses, dict)
+
+    # Test forward test
+    with torch.no_grad():
+        img_list = [img[None, :] for img in imgs]
+        for one_img in img_list:
+            recognizer(one_img, None, return_loss=False)
+
+    # Test forward gradcam
+    recognizer(imgs, gradcam=True)
+    for one_img in img_list:
+        recognizer(one_img, gradcam=True)
+
+
+def test_tsm():
+    model, train_cfg, test_cfg = get_recognizer_cfg(
+        'tsm/tsm_r50_1x1x8_50e_kinetics400_rgb.py')
+    model['backbone']['pretrained'] = None
+
+    recognizer = build_recognizer(
+        model, train_cfg=train_cfg, test_cfg=test_cfg)
+
+    input_shape = (1, 8, 3, 32, 32)
+    demo_inputs = generate_recognizer_demo_inputs(input_shape)
+
+    imgs = demo_inputs['imgs']
+    gt_labels = demo_inputs['gt_labels']
+
+    losses = recognizer(imgs, gt_labels)
+    assert isinstance(losses, dict)
+
+    # Test forward test
+    with torch.no_grad():
+        img_list = [img[None, :] for img in imgs]
+        for one_img in img_list:
+            recognizer(one_img, None, return_loss=False)
+
+    # test twice sample + 3 crops
+    input_shape = (2, 48, 3, 32, 32)
+    demo_inputs = generate_recognizer_demo_inputs(input_shape)
+    imgs = demo_inputs['imgs']
+
+    test_cfg = dict(average_clips='prob')
+    recognizer = build_recognizer(
+        model, train_cfg=train_cfg, test_cfg=test_cfg)
+
+    # Test forward test
+    with torch.no_grad():
+        img_list = [img[None, :] for img in imgs]
+        for one_img in img_list:
+            recognizer(one_img, None, return_loss=False)
+
+    # Test forward gradcam
+    recognizer(imgs, gradcam=True)
+    for one_img in img_list:
+        recognizer(one_img, gradcam=True)
